+++ apps/storybook/.storybook/walletConnector.tsx
@@ -1,0 +1,6 @@
+import { FC, ReactNode } from "react";
+import { WalletConnectorProvider } from "@orderly.network/wallet-connector";
+
+export const WalletConnector: FC<{ children: ReactNode }> = (props) => {
+  return <WalletConnectorProvider>{props.children}</WalletConnectorProvider>;
+};