+++ packages/types/src/types/api.ts
@@ -138,7 +138,7 @@ export declare namespace API {
     executed: number;
     total_fee: number;
     fee_asset: string;
-    client_order_id: null;
+    client_order_id?: string;
     average_executed_price: number;
     total_executed_quantity: number;
     visible_quantity: number;
@@ -179,6 +179,7 @@ export declare namespace API {
     created_time: number;
     updated_time: number;
     order_tag?: string;
+    client_order_id?: string;
   }
 
   export interface AlgoOrderExt extends AlgoOrder {
@@ -542,6 +543,18 @@ export declare namespace API {
     token: string;
     balance: number;
     pending_rebalance: number;
+  }
+
+  export interface RestrictedAreas {
+    invalid_web_country: string;
+    invalid_web_city: string;
+  }
+
+  export interface IpInfo {
+    ip: string;
+    city: string;
+    region: string;
+    checked: boolean;
   }
 }