+++ packages/referral/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @orderly.network/referral
 
+## 2.4.1
+
+### Patch Changes
+
+- publish
+- 5a220e0: publish
+- Updated dependencies
+- Updated dependencies [5a220e0]
+  - @orderly.network/react@2.4.1
+  - @orderly.network/hooks@2.4.1
+  - @orderly.network/utils@2.4.1
+
+## 2.4.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/react@2.4.1-alpha.0
+  - @orderly.network/hooks@2.4.1-alpha.0
+  - @orderly.network/utils@2.4.1-alpha.0
+
+## 2.4.0
+
+### Minor Changes
+
+- decb695: publish
+- publish
+
+### Patch Changes
+
+- Updated dependencies [decb695]
+- Updated dependencies
+  - @orderly.network/react@2.4.0
+  - @orderly.network/hooks@2.4.0
+  - @orderly.network/utils@2.4.0
+
+## 2.4.0-alpha.0
+
+### Minor Changes
+
+- publish
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/react@2.4.0-alpha.0
+  - @orderly.network/hooks@2.4.0-alpha.0
+  - @orderly.network/utils@2.4.0-alpha.0
+
 ## 2.3.2
 
 ### Patch Changes